+++ frontend/src/containers/map/content/map/popup/boundaries/index.tsx
@@ -2,7 +2,6 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import { useMap } from 'react-map-gl';
 
-import { useParams } from 'next/navigation';
 import { useRouter } from 'next/router';
 
 import type { Feature } from 'geojson';
@@ -10,7 +9,7 @@ import { useAtom, useAtomValue } from 'jotai';
 import { useLocale, useTranslations } from 'next-intl';
 
 import { PAGES } from '@/constants/pages';
-import { useMapSearchParams } from '@/containers/map/content/map/sync-settings';
+import { useMapSearchParams, useSyncMapLayers } from '@/containers/map/content/map/sync-settings';
 import { layersInteractiveIdsAtom, popupAtom } from '@/containers/map/store';
 import { formatPercentage, formatKM } from '@/lib/utils/formats';
 import { FCWithMessages } from '@/types';
@@ -19,35 +18,57 @@ import { useGetProtectionCoverageStats } from '@/types/generated/protection-cove
 import { ProtectionCoverageStat } from '@/types/generated/strapi.schemas';
 import { LayerTyped } from '@/types/layers';
 
-const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
+import { POPUP_BUTTON_CONTENT_BY_SOURCE, POPUP_PROPERTIES_BY_SOURCE } from '../constants';
+
+const BoundariesPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
   const t = useTranslations('containers.map');
   const locale = useLocale();
 
   const [rendered, setRendered] = useState(false);
-  const DATA_REF = useRef<Feature['properties'] | undefined>();
-  const { default: map } = useMap();
-  const searchParams = useMapSearchParams();
-  const { push } = useRouter();
-  const [popup, setPopup] = useAtom(popupAtom);
-  const { locationCode } = useParams();
 
+  const geometryDataRef = useRef<Feature['properties'] | undefined>();
+  const { default: map } = useMap();
+
+  const searchParams = useMapSearchParams();
+  const [activeLayers] = useSyncMapLayers();
+
+  const { push } = useRouter();
+
+  const [popup, setPopup] = useAtom(popupAtom);
   const layersInteractiveIds = useAtomValue(layersInteractiveIdsAtom);
 
-  const { data: source } = useGetLayersId<LayerTyped['config']['source']>(
+  const {
+    data: { source, environment },
+  } = useGetLayersId<{
+    source: LayerTyped['config']['source'];
+    environment: string;
+  }>(
     layerId,
     {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
+      locale,
       fields: ['config'],
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      populate: {
+        environment: {
+          fields: ['slug'],
+        },
+      },
     },
     {
       query: {
-        select: ({ data }) => (data.attributes as LayerTyped).config?.source,
+        placeholderData: { data: {} },
+        select: ({ data }) => ({
+          source: (data.attributes as LayerTyped)?.config?.source,
+          environment: data.attributes?.environment?.data.attributes.slug,
+        }),
       },
     }
   );
 
-  const DATA = useMemo(() => {
+  const geometryData = useMemo(() => {
     if (source?.type === 'vector' && rendered && popup && map) {
       const point = map.project(popup.lngLat);
 
@@ -58,7 +79,7 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
         point.y < 0 ||
         point.y > map.getCanvas().height
       ) {
-        return DATA_REF.current;
+        return geometryDataRef.current;
       }
       const query = map.queryRenderedFeatures(point, {
         layers: layersInteractiveIds,
@@ -68,23 +89,25 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
         return d.source === source.id;
       })?.properties;
 
-      DATA_REF.current = d;
+      geometryDataRef.current = d;
 
       if (d) {
-        return DATA_REF.current;
+        return geometryDataRef.current;
       }
     }
 
-    return DATA_REF.current;
+    return geometryDataRef.current;
   }, [popup, source, layersInteractiveIds, map, rendered]);
 
-  const selectedLocationCode: string = Object.keys(DATA || {})
-    .filter((key) => key.startsWith('ISO_') && DATA[key])
-    .map((key) => DATA[key])
-    .find((code, index, codes) => {
-      if (codes.length > 1) return code === locationCode;
-      return true;
-    });
+  const locationCode = useMemo(
+    () => geometryData?.[POPUP_PROPERTIES_BY_SOURCE[source?.['id']]?.id],
+    [geometryData, source]
+  );
+
+  const localizedLocationName = useMemo(
+    () => geometryData?.[POPUP_PROPERTIES_BY_SOURCE[source?.['id']]?.name[locale]],
+    [geometryData, locale, source]
+  );
 
   const { data: protectionCoverageStats, isFetching } =
     useGetProtectionCoverageStats<ProtectionCoverageStat>(
@@ -92,14 +115,14 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
         locale,
         filters: {
           location: {
-            code: selectedLocationCode,
+            code: locationCode,
           },
           is_last_year: {
             $eq: true,
           },
           environment: {
             slug: {
-              $eq: 'marine',
+              $eq: environment,
             },
           },
         },
@@ -107,7 +130,14 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
         // @ts-ignore
         populate: {
           location: {
-            fields: ['code', 'total_marine_area'],
+            fields: [
+              ...(locale === 'en' ? ['name'] : []),
+              ...(locale === 'es' ? ['name_es'] : []),
+              ...(locale === 'fr' ? ['name_fr'] : []),
+              'code',
+              'total_marine_area',
+              'total_terrestrial_area',
+            ],
           },
         },
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -117,42 +147,31 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
       },
       {
         query: {
-          enabled: !!selectedLocationCode,
-          select: ({ data }) => {
-            if (!data?.length) return undefined;
-            return data[0].attributes;
-          },
+          select: ({ data }) => data?.[0].attributes,
+          enabled: !!geometryData,
         },
       }
     );
 
-  const formattedCoverage = useMemo(() => {
-    if (protectionCoverageStats?.coverage !== undefined) {
-      return formatPercentage(locale, protectionCoverageStats.coverage, {
+  const formattedStats = useMemo(() => {
+    if (protectionCoverageStats) {
+      const percentage = formatPercentage(locale, protectionCoverageStats.coverage, {
         displayPercentageSign: false,
       });
+
+      const protectedArea = formatKM(locale, protectionCoverageStats.protected_area);
+
+      return {
+        percentage,
+        protectedArea,
+      };
     }
 
-    return '-';
+    return {
+      percentage: '-',
+      protectedArea: '-',
+    };
   }, [locale, protectionCoverageStats]);
-
-  const EEZName = useMemo(() => {
-    let name = null;
-
-    if (!DATA) {
-      return name;
-    }
-
-    if (locale === 'es') {
-      name = DATA.GEONAME_ES;
-    }
-
-    if (locale === 'fr') {
-      name = DATA.GEONAME_FR;
-    }
-
-    return name ?? DATA.GEONAME;
-  }, [locale, DATA]);
 
   // handle renderer
   const handleMapRender = useCallback(() => {
@@ -160,13 +179,9 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
   }, [map]);
 
   const handleLocationSelected = useCallback(async () => {
-    if (!protectionCoverageStats?.location?.data.attributes) return undefined;
-
-    const { code } = protectionCoverageStats.location.data.attributes;
-
-    await push(`${PAGES.progressTracker}/${code.toUpperCase()}?${searchParams.toString()}`);
+    await push(`${PAGES.progressTracker}/${locationCode.toUpperCase()}?${searchParams.toString()}`);
     setPopup({});
-  }, [push, searchParams, protectionCoverageStats, setPopup]);
+  }, [push, locationCode, searchParams, setPopup]);
 
   useEffect(() => {
     map?.on('render', handleMapRender);
@@ -178,11 +193,18 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
     };
   }, [map, handleMapRender]);
 
-  if (!DATA) return null;
+  // Close the tooltip if the layer that was clicked is not active anymore
+  useEffect(() => {
+    if (!activeLayers.includes(layerId)) {
+      setPopup({});
+    }
+  }, [layerId, activeLayers, setPopup]);
+
+  if (!geometryData) return null;
 
   return (
     <div className="space-y-2">
-      <h3 className="text-xl font-semibold">{EEZName}</h3>
+      <h3 className="text-xl font-semibold">{localizedLocationName || '-'}</h3>
       {isFetching && <div className="my-4 text-center font-mono text-xs">{t('loading')}</div>}
       {!isFetching && !protectionCoverageStats && (
         <div className="my-4 text-center font-mono text-xs">{t('no-data-available')}</div>
@@ -190,45 +212,53 @@ const EEZLayerPopup: FCWithMessages<{ layerId: number }> = ({ layerId }) => {
       {!isFetching && !!protectionCoverageStats && (
         <>
           <div className="space-y-2">
-            <div className="my-4 max-w-[95%] font-mono">{t('marine-conservation-coverage')}</div>
+            <div className="my-4 max-w-[95%] font-mono">
+              {environment === 'marine'
+                ? t('marine-conservation-coverage')
+                : t('terrestrial-conservation-coverage')}
+            </div>
             <div className="space-x-1 font-mono tracking-tighter text-black">
-              {formattedCoverage !== '-' &&
+              {formattedStats.percentage !== '-' &&
                 t.rich('percentage-bold', {
-                  percentage: formattedCoverage,
+                  percentage: formattedStats.percentage,
                   b1: (chunks) => (
                     <span className="text-[64px] font-bold leading-[80%]">{chunks}</span>
                   ),
                   b2: (chunks) => <span className="text-lg">{chunks}</span>,
                 })}
-              {formattedCoverage === '-' && (
-                <span className="text-[64px] font-bold leading-[80%]">{formattedCoverage}</span>
+              {formattedStats.percentage === '-' && (
+                <span className="text-[64px] font-bold leading-[80%]">
+                  {formattedStats.percentage}
+                </span>
               )}
             </div>
-            <div className="space-x-1 font-mono text-xs font-medium text-black">
+            <div className="space-x-1 font-mono font-medium text-black">
               {t('protected-area', {
-                protectedArea: formatKM(locale, protectionCoverageStats.protected_area),
+                protectedArea: formattedStats.protectedArea,
                 totalArea: formatKM(
                   locale,
-                  Number(protectionCoverageStats.location.data.attributes.total_marine_area)
+                  Number(
+                    protectionCoverageStats?.location.data.attributes[
+                      environment === 'marine' ? 'total_marine_area' : 'total_terrestrial_area'
+                    ]
+                  )
                 ),
               })}
             </div>
           </div>
-          {!!selectedLocationCode && (
-            <button
-              type="button"
-              className="block w-full border border-black p-4 text-center font-mono uppercase"
-              onClick={handleLocationSelected}
-            >
-              {t('open-country-insights')}
-            </button>
-          )}
+          <button
+            type="button"
+            className="block w-full border border-black p-4 text-center font-mono uppercase"
+            onClick={handleLocationSelected}
+          >
+            {t(POPUP_BUTTON_CONTENT_BY_SOURCE[source?.['id']])}
+          </button>
         </>
       )}
     </div>
   );
 };
 
-EEZLayerPopup.messages = ['containers.map'];
+BoundariesPopup.messages = ['containers.map'];
 
-export default EEZLayerPopup;
+export default BoundariesPopup;
